# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Advanced Talent Agency - Sports Streaming & Betting Exchange
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/suspend:
    patch:
      operationId: suspendUser
      tags: [users]
      summary: Suspend or unsuspend user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendUserInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # WALLET
...
  /streams:
    get:
      operationId: listStreams
      tags: [streams]
      summary: List streams
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Streams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamListResponse"
    post:
      operationId: createStream
      tags: [streams]
      summary: Create stream (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
...
  /streams/{id}:
    get:
      operationId: getStream
      tags: [streams]
      summary: Get stream by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
    patch:
      operationId: updateStream
      tags: [streams]
      summary: Update stream (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
    delete:
      operationId: deleteStream
      tags: [streams]
      summary: Delete stream (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /streams/{id}/access:
    post:
      operationId: purchaseStreamAccess
      tags: [streams]
      summary: Purchase 24h stream access ($1.50 deducted)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamAccess"
  /streams/{id}/access/check:
    get:
      operationId: checkStreamAccess
      tags: [streams]
      summary: Check if user has active stream access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Access status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamAccess"
...
  /streams/{id}/end:
    patch:
      operationId: endStream
      tags: [streams]
      summary: End stream (admin/moderator)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stream ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"

  # GAMES
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
...
    User:
      type: object
      required: [id, email, fullName, role, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [guest, user, moderator, admin]
        status:
          type: string
          enum: [active, suspended]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [user, moderator, admin]
    SuspendUserInput:
      type: object
      required: [suspended]
      properties:
        suspended:
          type: boolean
        reason:
          type: string
    UserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # WALLET
...
    Stream:
      type: object
      required: [id, title, sport, status, startTime, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        sport:
          type: string
          enum: [pool, boxing, other]
        thumbnailUrl:
          type: ["string", "null"]
        hlsUrl:
          type: ["string", "null"]
        streamKey:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, live, ended, cancelled]
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        viewerCount:
          type: integer
          default: 0
        accessPrice:
          type: number
          default: 1.5
        createdAt:
          type: string
...
    StreamInput:
      type: object
      required: [title, sport, startTime]
      properties:
        title:
          type: string
        description:
          type: string
        sport:
          type: string
          enum: [pool, boxing, other]
        thumbnailUrl:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        accessPrice:
          type: number
    StreamUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        sport:
          type: string
        thumbnailUrl:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
        hlsUrl:
          type: string
        accessPrice:
          type: number
...
    StreamAccess:
      type: object
      required: [hasAccess]
      properties:
        hasAccess:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        secondsRemaining:
          type: ["integer", "null"]
    StreamListResponse:
      type: object
      required: [streams, total, page, limit]
      properties:
        streams:
          type: array
          items:
            $ref: "#/components/schemas/Stream"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # GAMES
...